requirements.txt
# Core dependencies
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.5.0
django-cors-headers>=4.3.0,<5.0.0
django-environ>=0.11.2
gunicorn>=21.2.0,<22.0.0
whitenoise>=6.6.0,<7.0.0
python-dotenv>=1.0.0,<2.0.0

# Database
psycopg2-binary>=2.9.9,<3.0.0
GDAL==3.4.1  # Fixed version to match system

# Geo libraries
fastkml>=0.12,<0.13
geopy>=2.4.1,<3.0.0
pyproj>=3.6.1,<4.0.0
pygeoif>=0.6,<0.7

# Communication
paho-mqtt>=1.6.1,<2.0.0
pyserial>=3.5,<4.0.0
IMAPClient>=2.3.0,<3.0.0

# Utilities
Pillow>=10.0.0,<11.0.0
python-dateutil>=2.8.2,<3.0.0
pytz>=2024.1,<2025.0
requests>=2.31.0,<3.0.0
urllib3>=2.2.0,<3.0.0
certifi>=2024.2.2,<2025.0
chardet>=5.2.0,<6.0.0
idna>=3.6,<4.0.0

# Development tools
pytest>=7.4.0,<8.0.0
pytest-django>=4.7.0,<5.0.0
black>=24.1.0,<25.0.0
flake8>=7.0.0,<8.0.0
mypy>=1.8.0,<2.0.0

# Additional dependencies
redis>=5.0.1
django-redis>=5.4.0
Pillow>=10.1.0
django-filter>=23.5
django-storages>=1.14.2
boto3>=1.34.0
celery>=5.3.6
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
flower>=2.0.1 